chore: make State directory and move State files to State directory
