+++ Assets/SWPPT3/Scripts/Main/Player/Player.cs
@@ -2,6 +2,7 @@ using System.Collections.Generic;
 using Unity;
 using UnityEngine;
 using SWPPT3.Main.Prop;
+using SWPPT3.Main.Player.State;
 
 namespace SWPPT3.Main.Player
 {
