+++ Assets/SWPPT3/Scripts/Main/Player/State/PlayerState.cs
@@ -1,6 +1,6 @@
 using SWPPT3.Main.Prop;
 
-namespace SWPPT3.Main.Player
+namespace SWPPT3.Main.Player.State
 {
     public enum States
     {
