+++ Assets/SWPPT3/Scripts/Main/Prop/ItemBox.cs
@@ -1,4 +1,5 @@
 using SWPPT3.Main.Player;
+using SWPPT3.Main.Player.State;
 
 namespace SWPPT3.Main.Prop
 {
